fix(metadata): reuse fuzzy name match for Deezer artist enrichment
Deezer artist enrichment picked its hit with an exact case-insensitive
name equality — the same too-strict match that broke TheAudioDB bios in
#342. Deezer's search is accent-insensitive, so "Celine Dion" returned
"Céline Dion" but the equality filter dropped it, leaving the artist's
picture/cover/fans un-enriched (also missed "Bob Marley & The Wailers"
supersets and "Boney M." punctuation).

Extract the #342 matcher into a shared `waveflow_core::metadata::name_match`
module (`normalize_name` + generic `select_by_name`) and route both
TheAudioDB and Deezer through it. No behavior change for TheAudioDB — its
10 selection tests move to the shared module verbatim (via &str
candidates) and still pass.

Fixes #382
